Leave a clear line after spinner is stopped
Currently after spinner is stopped it might leave behind the last symbol
that was printed. This happens because '\b' is non-destructive in many
terminal emulators. If '\n' is printed after spinner is done - the last
printed symbol will remain printed (instead of an empty line). This
happens because '\n' also happens to be non-destructive.

VS Code is one example of such terminal emulator.